+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## v22.0.0
-
+## 22.0.0
+* add num_points_to_process to the config file to limit the points to be processed.
+* add starting_point to the config file to set the x,y,z location of the starting point for DVC analysis
 * Removed Eigen library from the source code. Eigen is downloaded from https://github.com/vais-ral/Eigen3 at build time.
 * Removed unused and old Python Wrapper code
 * Add strain executable to install target